Add generic `__repr__` and `to_dict` functions (#1143)
Adds a generic `__repr__` function and and a generic `to_dict` function
so that objects can be serialized in the future. The reason I didn't use
the SQLAlchemy class like we talked about in the previous PR was because
that would require a complete rework of most object instantiations and
how properties are ordered. For a greater view into that, you can look
at this PR: https://github.com/lucyparsons/OpenOversight/pull/1142

- [x] This branch is up-to-date with the `develop` branch.
- [x] `pytest` passes on my local development environment.
- [x] `pre-commit` passes on my local development environment.